move SEGraph to a separate compoenent
